pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "scripts"
version = "0.1.0"
description = "Helper scripts for the test-results project"
authors = [
    { name = "Benny Eisner", email = "benny.eisner@weather.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click",
    "requests",
    "lxml",
    "toml",
    "pydantic",
    "beautifulsoup4",
    "psycopg[binary]",  # v3 driver for PostgreSQL
    "python-dotenv",
    "Faker",
]

[project.optional-dependencies]
dev = [
    "autoflake",
    "invoke",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "coverage",
    "mypy",
    "flake8",
    "black",
    "isort",
    "mypy",
]

[project.scripts]
benny = "scripts.cli:cli"
